pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autobidsify"
version = "0.5.0"
description = "Automated BIDS standardization tool powered by LLM-first architecture"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Yiyi Liu", email = "yiyi.liu3@northeastern.edu"}
]
maintainers = [
    {name = "Yiyi Liu", email = "yiyi.liu3@northeastern.edu"}
]
keywords = [
    "bids",
    "neuroimaging",
    "mri",
    "fnirs",
    "dicom",
    "nifti",
    "data-standardization",
    "brain-imaging",
    "medical-imaging",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "pydicom>=2.4.0",
    "nibabel>=5.0.0",
    "h5py>=3.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=2.0",
]
dotenv = [
    "python-dotenv>=1.0.0",
]
all = [
    "autobidsify[dev,docs,dotenv]",
]

[project.urls]
Homepage = "https://github.com/fangzhouliucode/autobidsify"
Documentation = "https://autobidsify.readthedocs.io"
Repository = "https://github.com/fangzhouliucode/autobidsify"
Issues = "https://github.com/fangzhouliucode/autobidsify/issues"
Changelog = "https://github.com/fangzhouliucode/autobidsify/blob/main/CHANGELOG.md"

[project.scripts]
autobidsify = "autobidsify.__main__:main"
bidsify = "autobidsify.__main__:main"

# ============================================================================
# Build Configuration (FLAT LAYOUT)
# ============================================================================

[tool.hatch.build.targets.wheel]
packages = ["autobidsify"]

[tool.hatch.build.targets.sdist]
include = [
    "/autobidsify",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "PIE", "SIM", "RUF"]
ignore = ["E501", "B008", "C901", "PLR0913"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG001", "S101"]

[tool.ruff.isort]
known-first-party = ["autobidsify"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.coverage.run]
source = ["autobidsify"]
branch = true